requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
ultralytics==8.0.0
opencv-python==4.8.0.74
numpy==1.24.3
pillow==9.5.0
torch==2.0.0
torchvision==0.15.0